EOS-13239: Provide hare service to cluster managment
- Add hctl implementation
