--- plugin/service/hctl.py
@@ -1,23 +1,0 @@
-'''Wrapper module which provides hctl interface'''
-
-
-class HctlWrapper(ServicePlugin):
-    '''Wrapper class for hctl interface
-       implementation'''
-
-    def __init__(self):
-        ''''Init method'''
-        pass
-
-    def start(self):
-        '''starts the service using hctl'''
-        pass
-
-    def stop(self):
-        '''stops the service using hctl'''
-        pass
-
-    def status(self):
-        '''Returns the status of the service using
-           hctl interface'''
-        pass
